requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.2.0
psycopg==3.2.3
python-decouple==3.8
Pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
scikit-learn==1.5.1
nltk==3.9.1
django-cors-headers==4.3.1
google-generativeai==0.8.3
celery==5.3.4
redis==5.0.1
daphne==4.0.0
pandas
numpy
joblib

# API Documentation
drf-spectacular==0.27.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Performance testing
locust==2.17.0

# Secure file upload dependencies
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows compatibility
